% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliment-package.R
\docType{package}
\name{compliment-package}
\alias{compliment-package}
\title{compliment: Generate Randomized Compliments}
\description{
Generates randomized, uplifting compliments by combining adjectives and traits from a built-in word bank.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/guadapascal/compliment}
  \item Report bugs at \url{https://github.com/guadapascal/compliment/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
