% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliment.R
\name{compliment}
\alias{compliment}
\title{Generate a random compliment}
\usage{
compliment(n = 1)
}
\arguments{
\item{n}{Number of compliments to generate. Defaults to \code{1}.}
}
\value{
A character vector of length \code{n} containing compliments.
}
\description{
Randomly assembles a compliment from a bank of adjectives and traits.
}
\examples{
compliment()
compliment(3)

}
