% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farewell.R
\name{farewell}
\alias{farewell}
\title{Generate a random goodbye}
\usage{
farewell(n = 1)
}
\arguments{
\item{n}{Number of farewells to generate. Defaults to \code{1}.}
}
\value{
A character vector of length \code{n} containing farewells.
}
\description{
Randomly assembles a farewell from a bank of adjectives and closings.
}
\examples{
farewell()
farewell(3)

}
